backend/requirements.txt
fastapi
uvicorn
motor
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
python-multipart
email-validator
certifi
